re-order stops for better gallery flow
12 <-> 11
16 -> 13
13 -> 14
14 -> 16

TODO: rename the audio files as well! We aren't sure we actually want to do this yet, first we're printing new wall labels to see if that clears up the confusion. If that doesn't work we need to re-record the audio for the changed stops anywhere it mentions the old stop number